3. Интерфейс Input [#181092 #186078]

+++ src/main/java/ru/job4j/poly/Transport.java
@@ -5,5 +5,5 @@ public interface Transport {
 
     void passengers(int pass);
 
-    int refuel(int fuel);
+    double refuel(int fuel);
 }
